Rename auth to authorization consistently.
